Rename notify empty use case to alarm
